Make grouping and split accept double values
Make grouping and split accept double values was done for Stocks.

+++ app/src/main/java/br/com/guiainvestimento/adapter/stock/StockDetailAdapter.java
@@ -104,7 +104,7 @@ public class StockDetailAdapter extends RecyclerView.Adapter<RecyclerView.ViewHo
 
                 Long timestamp = mCursor.getLong(mCursor.getColumnIndex(PortfolioContract.StockTransaction.COLUMN_TIMESTAMP));
                 String date = TimestampToDate(timestamp);
-                int quantity = mCursor.getInt(mCursor.getColumnIndex(PortfolioContract.StockTransaction.COLUMN_QUANTITY));
+                double quantity = mCursor.getDouble(mCursor.getColumnIndex(PortfolioContract.StockTransaction.COLUMN_QUANTITY));
                 double price = mCursor.getDouble(mCursor.getColumnIndex(PortfolioContract.StockTransaction.COLUMN_PRICE));
                 // If price is 0, then it is bonification, grouping or split which should not show price or totalValue
                 if (price > 0) {
